+++ src/lib/components/PrimaryNameMark.tsx
@@ -3,8 +3,6 @@ import { Tooltip } from "lib/components/Tooltip";
 
 export const PrimaryNameMark = () => (
   <Tooltip label="Primary name">
-    <div>
-      <CustomIcon name="star-solid" color="accent.main" boxSize={3} m={0} />
-    </div>
+    <CustomIcon name="star-solid" color="accent.main" boxSize={3} m={0} />
   </Tooltip>
 );