+++ src/lib/pages/proposal-details/components/proposal-overview/status-summary/SummaryStatusBody.tsx
@@ -3,7 +3,7 @@ import { SkeletonText, Text } from "@chakra-ui/react";
 import big from "big.js";
 
 import type { ProposalOverviewProps } from "..";
-import { ErrorFetching } from "lib/components/state";
+import { ErrorFetchingProposalInfos } from "../../ErrorFetchingProposalInfos";
 import {
   extractParams,
   mapDeposit,
@@ -11,7 +11,11 @@ import {
 } from "lib/pages/proposal-details/utils";
 import type { Token, TokenWithValue, U } from "lib/types";
 import { ProposalStatus } from "lib/types";
-import { formatTokenWithValueList } from "lib/utils";
+import {
+  divWithDefault,
+  formatPrettyPercent,
+  formatTokenWithValueList,
+} from "lib/utils";
 
 const Passed = () => (
   <span
@@ -59,14 +63,16 @@ export const SummaryStatusBody = ({
 
   if (isLoading)
     return <SkeletonText mt={1} noOfLines={3} spacing={4} skeletonHeight={2} />;
-  if (!params || !votesInfo)
-    return <ErrorFetching dataName="proposal params and votes tally" />;
+  if (!params || !votesInfo) return <ErrorFetchingProposalInfos />;
 
   const { minDeposit, quorum, threshold, vetoThreshold } = extractParams(
     params,
     proposalData.isExpedited
   );
-  const { yes, noWithVeto, currentTotalVotes } = normalizeVotesInfo(votesInfo);
+  const { yes, noWithVeto, nonAbstainVotes, totalVotes } =
+    normalizeVotesInfo(votesInfo);
+  const yesRatio = divWithDefault(yes, nonAbstainVotes, 0);
+  const noWithVetoRatio = divWithDefault(noWithVeto, totalVotes, 0);
 
   if (proposalData.status === ProposalStatus.DEPOSIT_PERIOD) {
     const required = mapDeposit(proposalData.totalDeposit, minDeposit).reduce<
@@ -91,7 +97,7 @@ export const SummaryStatusBody = ({
   }
 
   if (proposalData.status === ProposalStatus.VOTING_PERIOD) {
-    if (currentTotalVotes.lt(quorum))
+    if (totalVotes.lt(quorum))
       return (
         <Text variant="body2">
           As of now, the proposal has not yet reached the required quorum. If
@@ -100,24 +106,40 @@ export const SummaryStatusBody = ({
         </Text>
       );
 
-    if (noWithVeto.gte(vetoThreshold))
+    if (noWithVetoRatio.gte(vetoThreshold))
       return (
         <Text variant="body2">
-          The proposal has successfully met the voting quorum. However, if the
-          voting period concludes with a{" "}
+          The proposal has{" "}
           <span
             style={{
               fontWeight: 700,
             }}
           >
-            &ldquo;No with veto&rdquo; vote surpassing{" "}
-            {Math.round(vetoThreshold * 10000) / 100}%
+            successfully met
+          </span>{" "}
+          the voting quorum. However, the &ldquo;No with veto&rdquo; vote
+          proportion currently constitutes{" "}
+          <span
+            style={{
+              fontWeight: 700,
+            }}
+          >
+            {formatPrettyPercent(noWithVetoRatio.toNumber())}
+          </span>{" "}
+          of the total votes, including &ldquo;Abstain&rdquo;, which exceeds the{" "}
+          <span
+            style={{
+              fontWeight: 700,
+            }}
+          >
+            {formatPrettyPercent(vetoThreshold)} threshold
           </span>
-          , the proposal will be <Rejected /> regardless of other votes.
+          . If the proposal concludes with this voting outcome, it will be
+          rejected regardless of &ldquo;Yes&rdquo; votes.
         </Text>
       );
 
-    if (currentTotalVotes.eq(0) || yes.div(currentTotalVotes).lt(threshold))
+    if (yesRatio.lt(threshold))
       return (
         <Text variant="body2">
           The proposal has{" "}
@@ -147,13 +169,13 @@ export const SummaryStatusBody = ({
     return (
       <Text variant="body2">
         Although the proposal successfully reached the voting quorum with a{" "}
-        {yes.mul(100).round(2, big.roundHalfUp).toNumber()}% &ldquo;Yes&rdquo;
-        rate, it was not implemented due to technical reasons.
+        {yesRatio.mul(100).round(2, big.roundHalfUp).toNumber()}%{" "}
+        &ldquo;Yes&rdquo; rate, it was not implemented due to technical reasons.
       </Text>
     );
 
   if (proposalData.status === ProposalStatus.REJECTED) {
-    if (currentTotalVotes.lt(quorum))
+    if (totalVotes.lt(quorum))
       return (
         <Text variant="body2">
           This proposal did not meet the required quorum, resulting in its
@@ -161,20 +183,26 @@ export const SummaryStatusBody = ({
         </Text>
       );
 
-    if (noWithVeto.gte(vetoThreshold))
+    if (noWithVetoRatio.gte(vetoThreshold))
       return (
         <Text variant="body2">
-          This proposal has{" "}
+          Due to the &ldquo;No with veto&rdquo; vote proportion constitutes{" "}
           <span
             style={{
               fontWeight: 700,
             }}
           >
-            reached
-          </span>{" "}
-          the voting quorum. But the voting period ended with &ldquo;No with
-          veto&rdquo; more than {Math.round(vetoThreshold * 10000) / 100}%, the
-          proposal will be <Rejected /> regardless of other votes.
+            {formatPrettyPercent(noWithVetoRatio.toNumber())}{" "}
+          </span>
+          of the total votes including &ldquo;Abstain&rdquo;, which exceeds the{" "}
+          <span
+            style={{
+              fontWeight: 700,
+            }}
+          >
+            {formatPrettyPercent(vetoThreshold)} threshold
+          </span>
+          , the proposal is <Rejected /> regardless of &ldquo;Yes&rdquo; votes.
         </Text>
       );