+++ src/lib/pages/proposal-details/data.ts
@@ -1,3 +1,4 @@
+import { useMobile } from "lib/app-provider";
 import { useAssetInfos } from "lib/services/assetService";
 import { useMovePoolInfos } from "lib/services/move";
 import {
@@ -11,12 +12,13 @@ export const useDerivedProposalParams = (): {
   data: Option<ProposalParams>;
   isLoading: boolean;
 } => {
+  const isMobile = useMobile();
   const { data, isLoading } = useProposalParams();
   const { data: assetInfos, isLoading: isAssetInfosLoading } = useAssetInfos({
-    withPrices: false,
+    withPrices: !isMobile,
   });
   const { data: movePoolInfos, isLoading: isMovePoolInfosLoading } =
-    useMovePoolInfos({ withPrices: false });
+    useMovePoolInfos({ withPrices: !isMobile });
 
   if (isLoading || isAssetInfosLoading || isMovePoolInfosLoading || !data)
     return {